Ready for deployment

+++ src/pages/api/message/[id].js
@@ -1,39 +1,34 @@
-import { dbConnect } from "utils/mongoose";
-dbConnect();
+import { dbConnect } from "utils/mongoose.js";
 import Message from "models/Message.model.js";
 import {isValidObjectId} from "mongoose";
+import { verifyAuth } from "dto/verifyAuth.js";
 
-
+dbConnect();
 
 
 export default async (req, res) => {
 
-
     const{method, query:{id}} = req;
-    
+
 
     const isId = isValidObjectId(id);
-
     if(!isId)return res.status(400).json("no valid id received");
 
 
+    const isLogin = await verifyAuth(req.headers.cookie);
 
     switch (method) {
 
-        case "GET":
-            const proj = await Message.findById(id);
-            if(!proj) return res.status(404).json({msg:"Message not found"});
-            return res.status(200).json(proj);
         case "DELETE":
-           
-            const deleteMsg = await Message.findByIdAndDelete(id);
-            if(!deleteMsg) return res.status(404).json({msg:"Msg not found"});
-            return res.status(204).json();
-        
+            if(isLogin){
+                const deleteMsg = await Message.findByIdAndDelete(id);
+                if(!deleteMsg) return res.status(404).json({msg:"msg not found"});
+                return res.status(200).json({iid:`${id}`});
+            }else{
+                return res.status(404).json({msg:"Not authorized"});
+            };
         default:
             return res.status(400).json({msg:"this method is not supported"});
-        
-       
     }
