+++ src/pages/api/post/[id].js
@@ -2,6 +2,7 @@
 import { dbConnect } from "utils/mongoose";
 import Post from "models/Post.model.js";
 import {isValidObjectId} from "mongoose"
+import { verifyAuth } from "dto/verifyAuth";
 
 dbConnect();
 
@@ -10,11 +11,14 @@ export default async (req, res) => {
 
     const{method, body, query:{id}} = req;
 
-    console.log(body, method)
 
     const isId = isValidObjectId(id);
-
     if(!isId)return res.status(400).json("no valid id received");
+
+
+    const isLogin = await verifyAuth(req.headers.cookie);
+
+
 
     switch (method) {
 
@@ -22,27 +26,31 @@ export default async (req, res) => {
             
             const post = await Post.findById(id);
             if(!post) return res.status(404).json({msg:"post not found"});
-            return res.status(200).json(post);
-
-
-           
+            return res.status(200).json(post);           
            
         case "PUT":
-            const upPost =  await Post.findByIdAndUpdate(id,body,{
-                new : true,
-            });
-            if(!upPost) return res.status(404).json({msg:"post not found"});
-            return res.status(201);
 
+            if(isLogin){
+                const upPost =  await Post.findByIdAndUpdate(id,body,{
+                    new : true,
+                });
+                if(!upPost) return res.status(404).json({msg:"post not found"});
+                return res.status(200);
+            }else{
+                return res.status(404).json({msg:"Not authorized"});
+            };
 
+           
 
         case "DELETE":
-            const deletePost = await Post.findByIdAndDelete(id);
-            if(!deletePost) return res.status(404).json({msg:"post not found"})
-            return res.status(204).json();
-        
-        
-        
+            
+            if(isLogin){
+                const deletePost = await Post.findByIdAndDelete(id);
+                if(!deletePost) return res.status(404).json({msg:"post not found"})
+                return res.status(200).json({iid:`${id}`});
+            }else{
+                return res.status(404).json({msg:"Not authorized"});
+            };
         default:
             return res.status(400).json({msg:"this method is not supported"})
     }
