+++ src/pages/api/project/[id].js
@@ -3,7 +3,7 @@ import { dbConnect } from "utils/mongoose";
 dbConnect();
 import Project from "models/Project.model.js";
 import {isValidObjectId} from "mongoose";
-
+import { verifyAuth } from "dto/verifyAuth";
 
 
 export default async (req, res) => {
@@ -11,8 +11,9 @@ export default async (req, res) => {
     const{method, body, query:{id}} = req;
     
     const isId = isValidObjectId(id);
-
     if(!isId)return res.status(400).json("no valid id received");
+
+    const isLogin = await verifyAuth(req.headers.cookie);
 
     switch (method) {
 
@@ -23,19 +24,31 @@ export default async (req, res) => {
            
            
         case "PUT":
-            const upProject =  await Project.findByIdAndUpdate(id,body,{
-                new : true,
-            });
-            if(!upProject) return res.status(404).json({msg:"Project not found"});
-            return res.status(200).json(upProject);
+
+            if (isLogin){
+                const upProject =  await Project.findByIdAndUpdate(id,body,{
+                    new : true,
+                });
+                if(!upProject) return res.status(404).json({msg:"Project not found"});
+                return res.status(200).json({msg:"editado"});
+            }else{
+                return res.status(404).json({msg:"Not authorized"});
+
+            };
+
+
 
 
 
         case "DELETE":
-
-            const deleteProject = await Project.findByIdAndDelete(id);
-            if(!deleteProject) return res.status(404).json({msg:"Project not found"});
-            return res.status(204).json();
+            if(isLogin){
+                const deleteProject = await Project.findByIdAndDelete(id);
+                if(!deleteProject) return res.status(404).json({msg:"Project not found"});
+                return res.status(200).json({iid:`${id}`});
+            }else{
+                return res.status(404).json({msg:"Not authorized"});
+            };
+           
         
         default:
             return res.status(400).json({msg:"this method is not supported"});
